Change HasFlags mixin to FlagCarryingRule superclass
The HasFlags mixin causes strange errors in FSH Online that are difficult (or possibly impossible) to debug. This abandons the mixin in favor of a common superclass, which is a reasonable solution to this problem.  At least for now.  Since we want to fix FSH Online *tonight*.
